HBAPP-33 Add prompt form component. Fix custom text input.
